% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scge.R
\docType{package}
\name{scge}
\alias{scge}
\alias{scge-package}
\title{scge: A package for simulating single cell gene expression data.}
\usage{
scge(data, meanObject = NULL, varObject = NULL, censorObject = NULL,
  copulaObject = NULL)
}
\arguments{
\item{data}{A matrix of gene expression counts. Rows should represent cells
and columns represent genes.}

\item{meanObject}{The option object from the \code{scgeMean} function.}

\item{varObject}{The option object from the \code{scgeVar} function.}

\item{censorObject}{The option object from the \code{scgeCensor} function.}

\item{copulaObject}{The option object from the \code{scgeCopula} function.}
}
\value{
\code{scge} returns an object of class "scge".
}
\description{
scge is a package to simulate realistic gene expression data, initially for
single cell genomics but generalisable to other sources of gene expression
data. Models are fitted to reproduce the mean expression, mean-variance,
mean-censorship and correlation relationships within the data. Parameters
from an example fit are provided should data be unavailable.

\code{scge} is used to fit gene expression models. This can be performed
automatically by passing a matrix a counts or by passing components of the
model.
}
\seealso{
\code{\link{scgeMean}}, \code{\link{scgeVar}},
\code{\link{scgeCensor}} and \code{\link{scgeCopula}}
}

