% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scgeCensor.R
\name{scgeCensor}
\alias{scgeCensor}
\title{Fitting Gene Censorship Models}
\usage{
scgeCensor(data = NULL)
}
\arguments{
\item{data}{A matrix of gene expression counts. Rows should represent cells
and columns represent genes.}
}
\value{
\code{scgeCensor} returns an object of class "scgeCensor".
}
\description{
\code{scge} is used to fit gene censorship models.
}
\seealso{
\code{\link{scge}}, \code{\link{scgeMean}}, \code{\link{scgeVar}}
and \code{\link{scgeCopula}}
}

