% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scgeMean.R
\name{scgeMean}
\alias{scgeMean}
\title{Fitting Gene Mean Expression Models}
\usage{
scgeMean(data = NULL)
}
\arguments{
\item{data}{A matrix of gene expression counts. Rows should represent cells
and columns represent genes.}
}
\value{
\code{scgeMean} returns an object of class "scgeMean".
}
\description{
\code{scge} is used to fit gene mean expression models.
}
\seealso{
\code{\link{scge}}, \code{\link{scgeVar}}, \code{\link{scgeCensor}}
and \code{\link{scgeCopula}}
}

