% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scgeVar.R
\name{scgeVar}
\alias{scgeVar}
\title{Fitting Gene Mean-Variance Models}
\usage{
scgeVar(data)
}
\arguments{
\item{data}{A matrix of gene expression counts. Rows should represent cells
and columns represent genes.}
}
\value{
\code{scgeVar} returns an object of class "scgeVar".
}
\description{
\code{scge} is used to fit gene mean-variance models.
}
\seealso{
\code{\link{scge}}, \code{\link{scgeMean}}, \code{\link{scgeCensor}}
and \code{\link{scgeCopula}}
}

