% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scgeCopula.R
\name{scgeCopula}
\alias{scgeCopula}
\title{Fitting Gene Correlation Models}
\usage{
scgeCopula(data)
}
\arguments{
\item{data}{A matrix of gene expression counts. Rows should represent cells
and columns represent genes.}
}
\value{
\code{scgeCopula} returns an object of class "scgeCopula".
}
\description{
\code{scge} is used to fit gene correlation models.
}
\seealso{
\code{\link{scge}}, \code{\link{scgeMean}}, \code{\link{scgeVar}}
and \code{\link{scgeCensor}}
}

